+++ previous-versions/commerce/src/sdk/models/operations/getproduct.ts
@@ -55,5 +55,5 @@ export class GetProductResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }