+++ previous-versions/commerce/src/sdk/models/operations/listorders.ts
@@ -73,5 +73,5 @@ export class ListOrdersResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }