+++ previous-versions/commerce/src/sdk/models/operations/listtaxcomponents.ts
@@ -67,7 +67,7 @@ export class ListTaxComponentsResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * OK