+++ previous-versions/commerce/src/sdk/sdk.ts
@@ -62,9 +62,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.39.0";
-    genVersion = "2.188.1";
-    userAgent = "speakeasy-sdk/typescript 0.39.0 2.188.1 3.0.0 @codat/commerce";
+    sdkVersion = "0.39.1";
+    genVersion = "2.194.1";
+    userAgent = "speakeasy-sdk/typescript 0.39.1 2.194.1 3.0.0 @codat/commerce";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);