+++ previous-versions/commerce/RELEASES.md
@@ -669,3 +669,13 @@ Based on:
 - [typescript v0.39.0] previous-versions/commerce
 ### Releases
 - [NPM v0.39.0] https://www.npmjs.com/package/@codat/commerce/v/0.39.0 - previous-versions/commerce
+
+## 2023-11-20 01:29:29
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Commerce.yaml
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.39.1] previous-versions/commerce
+### Releases
+- [NPM v0.39.1] https://www.npmjs.com/package/@codat/commerce/v/0.39.1 - previous-versions/commerce